+++ dart/geobase/CHANGELOG.md
@@ -10,6 +10,7 @@ NOTE: Version 0.6.0 currently [under development](https://github.com/navibyte/ge
 * [Redesigned PositionData supporting positions from coordinate value array and position objects #200](https://github.com/navibyte/geospatial/issues/200)
 * [Coordinates using data structures from typed data #203](https://github.com/navibyte/geospatial/issues/203)
 * [Position from subview of coordinate value array as doubles #204](https://github.com/navibyte/geospatial/issues/204)
+* [PositionSeries and geometries with reversed positions #205](https://github.com/navibyte/geospatial/issues/205)
 
 ⚠️ Breaking changes:
 * [Allow a position stored as "Position" in Point, and a bounding box as "Box" in Geometry classes #192](https://github.com/navibyte/geospatial/issues/192)